gorou ascension refactor
